requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
pip==22.2.2
pytz==2022.6
requests==2.28.1
setuptools==65.5.0
sqlparse==0.4.1
tzdata==2021.1
uritemplate==4.1.1
urllib3==1.26.12
wheel==0.37.1
wincertstore==0.2
